Refactor Zalmoxis model to include material dictionaries in main function and related calculations
